Improved readability in the adapters, and copied to the parent adapters folder
