+++ mediation.adaters/net/pubnative/mediation/adapter/network/YahooNetworkAdapter.java
@@ -23,151 +23,19 @@
 
 package net.pubnative.mediation.adapter.network;
 
-import android.content.Context;
-import android.text.TextUtils;
-import android.util.Log;
+import net.pubnative.mediation.adapter.PubnativeNetworkHub;
 
-import com.flurry.android.FlurryAgent;
-import com.flurry.android.ads.FlurryAdErrorType;
-import com.flurry.android.ads.FlurryAdNative;
-import com.flurry.android.ads.FlurryAdNativeListener;
+public class YahooNetworkAdapter extends PubnativeNetworkHub {
 
-import net.pubnative.mediation.adapter.PubnativeNetworkRequestAdapter;
-import net.pubnative.mediation.adapter.model.FlurryNativeAdModel;
-import net.pubnative.mediation.exceptions.PubnativeException;
+    @Override
+    public PubnativeNetworkRequestAdapter getRequestAdapter() {
 
-import java.util.Map;
-
-public class YahooNetworkAdapter extends PubnativeNetworkRequestAdapter implements FlurryAdNativeListener {
-
-    private static      String TAG                = YahooNetworkAdapter.class.getSimpleName();
-    public static final String KEY_AD_SPACE_NAME  = "ad_space_name";
-    public static final String KEY_FLURRY_API_KEY = "api_key";
-    protected Context mContext;
-
-    public YahooNetworkAdapter(Map data) {
-
-        super(data);
-    }
-
-    //==============================================================================================
-    // PubnativeNetworkAdapter
-    //==============================================================================================
-    public void start(Context context) {
-
-        Log.v(TAG, "start");
-        if (context == null || mData == null) {
-            invokeFailed(PubnativeException.ADAPTER_ILLEGAL_ARGUMENTS);
-        } else {
-            String apiKey = (String) mData.get(KEY_FLURRY_API_KEY);
-            String adSpaceName = (String) mData.get(KEY_AD_SPACE_NAME);
-            if (TextUtils.isEmpty(apiKey) || TextUtils.isEmpty(adSpaceName)) {
-                invokeFailed(PubnativeException.ADAPTER_MISSING_DATA);
-            } else {
-                mContext = context;
-                createRequest(adSpaceName, apiKey);
-            }
-        }
-    }
-
-    //==============================================================================================
-    // YahooNetworkAdapter
-    //==============================================================================================
-    protected void createRequest(String adSpaceName, String apiKey) {
-
-        Log.v(TAG, "createRequest");
-        // configure flurry
-        FlurryAgent.setLogEnabled(true);
-        // initialize flurry with new apiKey
-        FlurryAgent.init(mContext, apiKey);
-        // start/resume session
-        if (!FlurryAgent.isSessionActive()) {
-            FlurryAgent.onStartSession(mContext);
-        }
-        // Make request
-        FlurryAdNative flurryAdNative = new FlurryAdNative(mContext, adSpaceName);
-        flurryAdNative.setListener(this);
-        flurryAdNative.fetchAd();
-    }
-    //==============================================================================================
-    // Callbacks
-    //==============================================================================================
-
-    // FlurryAdNativeListener
-    //----------------------------------------------------------------------------------------------
-    protected void endFlurrySession(Context context) {
-
-        Log.v(TAG, "endFlurrySession");
-        FlurryAgent.onEndSession(context);
+        return new YahooNetworkRequestAdapter(mNetworkData);
     }
 
     @Override
-    public void onFetched(FlurryAdNative flurryAdNative) {
+    public PubnativeNetworkInterstitialAdapter getInterstitialAdapter() {
 
-        Log.v(TAG, "onFetched");
-        endFlurrySession(mContext);
-        FlurryNativeAdModel nativeAdModel = new FlurryNativeAdModel(flurryAdNative);
-        invokeLoaded(nativeAdModel);
-    }
-
-    @Override
-    public void onError(FlurryAdNative flurryAdNative, FlurryAdErrorType flurryAdErrorType, int errCode) {
-
-        Log.v(TAG, "onError: " + errCode);
-        endFlurrySession(mContext);
-        if (flurryAdErrorType == null) {
-            invokeFailed(PubnativeException.ADAPTER_UNKNOWN_ERROR);
-        } else if (FlurryAdErrorType.FETCH == flurryAdErrorType) {
-            invokeLoaded(null);
-        } else {
-            invokeFailed(new Exception("YahooNetworkAdapter - " + errCode + " - " + flurryAdErrorType.name()));
-        }
-    }
-
-    @Override
-    public void onShowFullscreen(FlurryAdNative flurryAdNative) {
-
-        Log.v(TAG, "onShowFullscreen");
-        // Do nothing for now.
-    }
-
-    @Override
-    public void onCloseFullscreen(FlurryAdNative flurryAdNative) {
-
-        Log.v(TAG, "onCloseFullscreen");
-        // Do nothing for now.
-    }
-
-    @Override
-    public void onAppExit(FlurryAdNative flurryAdNative) {
-
-        Log.v(TAG, "onAppExit");
-        // Do nothing for now.
-    }
-
-    @Override
-    public void onClicked(FlurryAdNative flurryAdNative) {
-
-        Log.v(TAG, "onClicked");
-        // Do nothing for now.
-    }
-
-    @Override
-    public void onImpressionLogged(FlurryAdNative flurryAdNative) {
-
-        Log.v(TAG, "onImpressionLogged");
-        // Do nothing for now.
-    }
-
-    @Override
-    public void onExpanded(FlurryAdNative flurryAdNative) {
-
-        Log.v(TAG, "onExpanded");
-    }
-
-    @Override
-    public void onCollapsed(FlurryAdNative flurryAdNative) {
-
-        Log.v(TAG, "onCollapsed");
+        return null;
     }
 }
