+++ mediation.demo/src/main/java/net/pubnative/mediation/demo/AdViewHolder.java
@@ -68,7 +68,7 @@ public class AdViewHolder implements PubnativeNetworkRequest.Listener,
 
     public AdViewHolder(Context context, View convertView) {
 
-        this.mContext = context;
+        mContext = context;
         mAdLoading = (ProgressBar) convertView.findViewById(R.id.ad_spinner);
         mAdContainer = (RelativeLayout) convertView.findViewById(R.id.ad_clickable);
         mRequestButton = (Button) convertView.findViewById(R.id.request_button);
@@ -125,7 +125,7 @@ public class AdViewHolder implements PubnativeNetworkRequest.Listener,
             mRating.setVisibility(View.VISIBLE);
             Picasso.with(mContext).load(model.getIconUrl()).into(mIcon);
             Picasso.with(mContext).load(model.getBannerUrl()).into(mBanner);
-            View sponsorView = model.getAdvertisingDisclosureView(this.mContext);
+            View sponsorView = model.getAdvertisingDisclosureView(mContext);
             if (sponsorView != null) {
                 mAdDisclosure.addView(sponsorView);
             }
@@ -147,12 +147,6 @@ public class AdViewHolder implements PubnativeNetworkRequest.Listener,
     //==============================================================================================
     // PubnativeNetworkRequest.Listener
     //----------------------------------------------------------------------------------------------
-    @Override
-    public void onPubnativeNetworkRequestStarted(PubnativeNetworkRequest request) {
-
-        Log.d(TAG, "onPubnativeNetworkRequestStarted");
-    }
-
     @Override
     public void onPubnativeNetworkRequestLoaded(PubnativeNetworkRequest request, PubnativeAdModel ad) {
